# fix(crypto): move FNV hashes from JS to Rust/Wasm, fixing iterator behaviour (#3811) and performance