--- a/crypto/_benches/bench.ts
+++ b/crypto/_benches/bench.ts
@@ -1,67 +1,70 @@
-#!/usr/bin/env -S deno run
+#!/usr/bin/env -S deno bench
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
-import { assert, assertEquals } from "../../assert/mod.ts";
+import { crypto as stdCrypto, DIGEST_ALGORITHM_NAMES } from "../mod.ts";
 
-import { crypto as stdCrypto } from "../mod.ts";
+import { crypto as oldCrypto } from "https://deno.land/std@0.220.1/crypto/mod.ts";
 
 const webCrypto = globalThis.crypto;
 
-// Wasm is limited to 32-bit operations, which SHA-256 is optimized for, while
-// SHA-512 is optimized for 64-bit operations and may be slower.
-for (const algorithm of ["SHA-256", "SHA-512"] as const) {
-  for (
-    const length of [
-      64,
-      262_144,
-      4_194_304,
-      67_108_864,
-      524_291_328,
-    ] as const
-  ) {
-    // Create a test input buffer and do some operations to hopefully ensure
-    // it's fully initialized and not optimized away.
-    const buffer = new Uint8Array(length);
-    for (let i = 0; i < length; i++) {
-      buffer[i] = (i + (i % 13) + (i % 31)) % 255;
-    }
-    let sum = 0;
-    for (const byte of buffer) {
-      sum += byte;
-    }
-    assert(sum > 0);
+const BENCHMARKED_DIGEST_ALGORITHM_NAMES = [
+  "SHA-256",
+  "SHA-512",
+  "BLAKE3",
+  "FNV32A",
+  "FNV64A",
+] satisfies (typeof DIGEST_ALGORITHM_NAMES[number])[];
 
-    for (
-      const implementation of [
-        "runtime WebCrypto (target)",
-        "std/crypto Wasm   (you are here)",
-      ] as const
+const WEB_CRYPTO_DIGEST_ALGORITHM_NAMES = [
+  "SHA-1",
+  "SHA-256",
+  "SHA-384",
+  "SHA-512",
+] satisfies (typeof DIGEST_ALGORITHM_NAMES[number])[];
+
+for (
+  const [length, humanLength] of [
+    [64, "64 B"],
+    [262_144, "256 KiB"],
+    [4_194_304, "4 MiB"],
+    [67_108_864, "64 MiB"],
+    [524_291_328, "512 MiB"],
+  ] as const
+) {
+  const buffer = new Uint8Array(length);
+  for (let i = 0; i < length; i++) {
+    buffer[i] = (i + (i % 13) + (i % 31)) % 256;
+  }
+
+  for (const name of ["FNV32A", "FNV64A"] as const) {
+    Deno.bench({
+      group: `${humanLength} in ${name}`,
+      name: `TypeScript (from v0.220.1) with ${humanLength} in ${name}`,
+      async fn() {
+        await oldCrypto.subtle.digest(name, buffer);
+      },
+    });
+  }
+
+  for (const name of BENCHMARKED_DIGEST_ALGORITHM_NAMES) {
+    if (
+      (WEB_CRYPTO_DIGEST_ALGORITHM_NAMES as readonly string[]).includes(name)
     ) {
-      let lastDigest: ArrayBuffer | undefined;
-
       Deno.bench({
-        name: `${algorithm.padEnd(12)} ${
-          length
-            .toString()
-            .padStart(12)
-        }B ${implementation}`,
+        group: `${humanLength} in ${name}`,
+        name: `Runtime WebCrypto with ${humanLength} in ${name}`,
         async fn() {
-          let digest;
-          if (implementation === "std/crypto Wasm   (you are here)") {
-            digest = stdCrypto.subtle.digestSync(algorithm, buffer);
-          } else if (implementation === "runtime WebCrypto (target)") {
-            digest = await webCrypto.subtle.digest(algorithm, buffer);
-          } else {
-            throw new Error(`Unknown implementation ${implementation}`);
-          }
-
-          assert(digest.byteLength > 0);
-
-          if (lastDigest) {
-            assertEquals(lastDigest, digest);
-          }
-          lastDigest = digest;
+          await webCrypto.subtle.digest(name, buffer);
         },
       });
     }
+
+    Deno.bench({
+      group: `${humanLength} in ${name}`,
+      name: `Rust/Wasm with ${humanLength} in ${name}`,
+      baseline: true,
+      async fn() {
+        await stdCrypto.subtle.digest(name, [buffer]);
+      },
+    });
   }
 }