The current implementation of the FNV hashing functions in std/crypto is implemented in TypeScript, unlike the other (cryptographic) hashing functions which are implemented in Rust/Wasm. [The intent of this choice](https://github.com/denoland/deno_std/issues/2122#issuecomment-1114242933) was to avoid the overhead of calling into Wasm, and to avoid the need of depending on third-party code for an algorithm that's simple enough to implement ourselves and for which we don't need to make any cryptographic guarantees.

Avoiding Wasm does yield performance benefits in some cases. On my machine, FNV32 hashing 64 bytes of data typically takes about 2 microseconds with either the TypeScript or Wasm implementations, but initializing the Wasm module adds a one-time cost of about 2 milliseconds and the Wasm performance is higher-variance, at least while it's warming up (with a 99th percentile performance of 10 microseconds). So if you're hashing a lot of very small values, the TypeScript implementation can come out ahead.

However, in most other cases the situation is different. The TypeScript FNV64 implementation incurs a lot of extra overhead from how it performs 64-bit integer operations using JavaScript's 64-bit float number type. Hashing 64 bytes of data takes 8 microseconds in TypeScript versus 2 microseconds in Wasm (4x slower, after the one-time initialization cost). If the data becomes much larger, the TypeScript FNV performance abruptly goes off a cliff:

_(Numbers slightly rounded for simplicity, using the benchmark in `crypto/_benches/bench.ts`.)_

- 64 B in FNV32: 2 microseconds in Wasm versus 2 microseconds in TypeScript (roughly equal)
- 256 KiB in FNV32: 300 microseconds in Wasm versus 6 milliseconds in TypeScript (~20x slower)
- 4 MiB in FNV32: 4 milliseconds in Wasm versus 100 milliseconds in TypeScript (~25x slower)
- 64 MiB in FNV32: 75 milliseconds in Wasm versus 1.6 seconds in TypeScript (~20x slower)
- 512 MiB in FNV32: 550 milliseconds in Wasm versus 13 seconds in TypeScript (~20x slower)

<!-- -->

- 64 B in FNV64: 2 microseconds in Wasm versus 8 microseconds in TypeScript (~4x slower)
- 256 KiB in FNV64: 300 microseconds in Wasm versus 25 milliseconds in TypeScript (~80x slower)
- 4 MiB in FNV64: 4 milliseconds in Wasm versus 375 milliseconds in TypeScript (~90x slower)
- 64 MiB in FNV64: 75 milliseconds in Wasm versus 5.9 seconds in TypeScript (~80x slower)
- 512 MiB in FNV64: 575 milliseconds in Wasm versus 45 seconds in TypeScript (~80x slower)

This means that for most cases, using the current FNV implementation is slower than using one of the Wasm cryptographic hash implementations, when the reason it exists is supposed to be to provide a faster alternative for non-cryptographic use cases. This is an issue.

Maybe it's possible to optimize the TypeScript implementation to close this gap, but I'm not sure how to do that, while I do know how to move the code into our Rust/Wasm bundle, so that's what this PR does.

The TypeScript FNV implementation has the separate issue that it currently doesn't support iterable inputs, such as Web Streams, but that's something our digest functions are meant to support and the Wasm/Rust implementations do (this PR resolves #3811). It can be refactored to handle this case, but when I did so that resulted in some duplication of logic and additional complexity that goes away when we move it into Rust/Wasm instead.

This change increases the size of `deno_std_wasm_crypto.generated.mjs` from 194,324B to 196,614B (a 1% increase of 2,290B), or from 77,230B to 78,068B after `gzip --best` (a 1% increase of 838B). (For some context, this PR follows #4510 which reduces those initial sizes by 18% and 13% respectively.)

There are several existing implementations of FNV hashes on crates.io, however none of the widely-used ones provide all four  variants we need. Since it's only a handful of lines of logic, I think it's safer just to implement it ourselves rather than add a new dependency on a not-widely-used crate, so that's what I've done here.

This PR merges the tests for FNV hashes with our tests for other hash functions, and adds the six test inputs specified in [IETF draft-eastlake-fnv-21](https://www.ietf.org/archive/id/draft-eastlake-fnv-21.html#name-a-few-test-vectors) and the four from [Go's standard library tests](https://github.com/golang/go/blob/go1.22.1/src/hash/fnv/fnv_test.go#L16-L48) to validate our implementation.

The PR updates the benchmark script to also include results for a previously-published version of this library, to provide a baseline for changes.

BREAKING CHANGE: This PR also renames the exported `wasmDigestAlgorithms` to `DIGEST_ALGORITHM_NAMES`, to match Deno's style guide and reflect the fact that it includes all algorithm names supported by the digest function, not just a subset. The exported `WasmDigestAlgorithm` type is removed, as it's now redundant with the existing `DigestAlgorithmName` export. (Generally, the fact that some algorithms are implemented in Wasm should probably be considered an implementation detail, and not exposed in the public interface.) The `FNVAlgorithms` export is also removed, since they're no longer a special case.