@jeremyBanks 2024-03-23
(marked as draft because it includes/depends on changes from https://github.com/denoland/deno_std/pull/4510, but this is mostly finished)