@jeremyBanks 2024-03-23
Here's my latest full output of `deno bench crypto --filter=FNV`. I'm not sure why the Wasm implementation has very high variance for FNV32A, while for FNV64A it has low variance and the TypeScript implementation has very high variance instead. Maybe it's something related to JIT warm-up, but I'm just guessing. However, I think the TypeScript implementation's extremely poor performance with non-tiny data is an overriding concern regardless of what's happening there.

<details><summary>As Text</summary>

```default
cpu: Intel(R) Core(TM) i7-7700 CPU @ 3.60GHz
runtime: deno 1.41.3 (x86_64-unknown-linux-gnu)

file:///mnt/c/Users/_/denoland/deno_std/crypto/_benches/bench.ts
benchmark                                              time (avg)        iter/s             (min … max)       p75       p99      p995
------------------------------------------------------------------------------------------------------- -----------------------------

TypeScript (from v0.220.1) with 64 B in FNV32A          2.29 µs/iter     436,906.1     (2.01 µs … 3.47 µs) 2.44 µs 3.47 µs 3.47 µs
Rust/Wasm with 64 B in FNV32A                           3.01 µs/iter     332,446.8       (2.2 µs … 2.4 ms) 2.6 µs 15 µs 16.9 µs

summary
  Rust/Wasm with 64 B in FNV32A
   1.31x slower than TypeScript (from v0.220.1) with 64 B in FNV32A

group 64 B in FNV64A
TypeScript (from v0.220.1) with 64 B in FNV64A          6.71 µs/iter     149,009.1     (5.4 µs … 967.7 µs) 6.1 µs 21.7 µs 26.6 µs
Rust/Wasm with 64 B in FNV64A                           1.44 µs/iter     696,684.8     (1.38 µs … 1.69 µs) 1.43 µs 1.69 µs 1.69 µs

summary
  Rust/Wasm with 64 B in FNV64A
   4.68x faster than TypeScript (from v0.220.1) with 64 B in FNV64A

group 256 KiB in FNV32A
TypeScript (from v0.220.1) with 256 KiB in FNV32A       6.36 ms/iter         157.2    (5.54 ms … 10.42 ms) 6.58 ms 10.42 ms 10.42 ms
Rust/Wasm with 256 KiB in FNV32A                      295.27 µs/iter       3,386.7     (273.4 µs … 478 µs) 299.5 µs 413.8 µs 448.9 µs

summary
  Rust/Wasm with 256 KiB in FNV32A
   21.54x faster than TypeScript (from v0.220.1) with 256 KiB in FNV32A

group 256 KiB in FNV64A
TypeScript (from v0.220.1) with 256 KiB in FNV64A       22.2 ms/iter          45.1   (21.38 ms … 24.89 ms) 22.24 ms 24.89 ms 24.89 ms
Rust/Wasm with 256 KiB in FNV64A                      315.01 µs/iter       3,174.5   (272.9 µs … 591.8 µs) 324.7 µs 450.4 µs 472.5 µs

summary
  Rust/Wasm with 256 KiB in FNV64A
   70.46x faster than TypeScript (from v0.220.1) with 256 KiB in FNV64A

group 4 MiB in FNV32A
TypeScript (from v0.220.1) with 4 MiB in FNV32A        98.95 ms/iter          10.1  (95.66 ms … 104.81 ms) 100.12 ms 104.81 ms 104.81 ms
Rust/Wasm with 4 MiB in FNV32A                          5.15 ms/iter         194.2     (4.54 ms … 6.15 ms) 5.43 ms 6.14 ms 6.15 ms

summary
  Rust/Wasm with 4 MiB in FNV32A
   19.22x faster than TypeScript (from v0.220.1) with 4 MiB in FNV32A

group 4 MiB in FNV64A
TypeScript (from v0.220.1) with 4 MiB in FNV64A        376.1 ms/iter           2.7 (353.45 ms … 444.37 ms) 378.96 ms 444.37 ms 444.37 ms
Rust/Wasm with 4 MiB in FNV64A                          4.77 ms/iter         209.6      (4.56 ms … 5.6 ms) 4.8 ms 5.32 ms 5.6 ms

summary
  Rust/Wasm with 4 MiB in FNV64A
   78.83x faster than TypeScript (from v0.220.1) with 4 MiB in FNV64A

group 64 MiB in FNV32A
TypeScript (from v0.220.1) with 64 MiB in FNV32A         1.74 s/iter           0.6       (1.62 s … 1.92 s) 1.78 s 1.92 s 1.92 s
Rust/Wasm with 64 MiB in FNV32A                         80.7 ms/iter          12.4   (76.25 ms … 85.21 ms) 83.29 ms 85.21 ms 85.21 ms

summary
  Rust/Wasm with 64 MiB in FNV32A
   21.57x faster than TypeScript (from v0.220.1) with 64 MiB in FNV32A

group 64 MiB in FNV64A
TypeScript (from v0.220.1) with 64 MiB in FNV64A         6.02 s/iter           0.2       (5.88 s … 6.26 s) 6.07 s 6.26 s 6.26 s
Rust/Wasm with 64 MiB in FNV64A                        81.36 ms/iter          12.3   (76.25 ms … 88.74 ms) 82.67 ms 88.74 ms 88.74 ms

summary
  Rust/Wasm with 64 MiB in FNV64A
   73.98x faster than TypeScript (from v0.220.1) with 64 MiB in FNV64A

group 512 MiB in FNV32A
TypeScript (from v0.220.1) with 512 MiB in FNV32A       13.46 s/iter           0.1      (12.99 s … 14.4 s) 13.93 s 14.4 s 14.4 s
Rust/Wasm with 512 MiB in FNV32A                      618.55 ms/iter           1.6  (609.8 ms … 626.73 ms) 621.93 ms 626.73 ms 626.73 ms

summary
  Rust/Wasm with 512 MiB in FNV32A
   21.75x faster than TypeScript (from v0.220.1) with 512 MiB in FNV32A

group 512 MiB in FNV64A
TypeScript (from v0.220.1) with 512 MiB in FNV64A       46.07 s/iter           0.0       (45.3 s … 48.5 s) 46.76 s 48.5 s 48.5 s
Rust/Wasm with 512 MiB in FNV64A                      597.24 ms/iter           1.7 (585.13 ms … 610.93 ms) 609.49 ms 610.93 ms 610.93 ms

summary
  Rust/Wasm with 512 MiB in FNV64A
   77.14x faster than TypeScript (from v0.220.1) with 512 MiB in FNV64A
```

</details>

<details><summary>As Image</summary>
<img width="626" alt="image" src="https://github.com/denoland/deno_std/assets/18020/c6291827-5fe1-412b-af8c-e825224f2ca5">
</details>

If anyone wants to try the benchmark on a machine with more stable performance than mine, please share the results!